[#64] move place-search service on google-sdk module & remove search-place module
