+++ app/components/places/pin/pin-place.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit, ViewChild, ElementRef } from "@angular/core";
 import { LoadingIndicator } from "nativescript-loading-indicator";
 
 import { PlacePinSearchComponent } from './search/place-pin-search.component';
-import { PlaceSearchService } from '../../../shared/place/search/place-search.service';
+import { PlaceSearchService } from '../../../modules/google-sdk/shared/place/search/place-search.service';
 import { GeolocationService } from '../../../shared/geolocation/geolocation.sercice';
 import providePlaceStorageService from '../../../shared/place/storage/provide-service';
 import { PlaceStorageService } from '../../../shared/place/storage/place-storage.service';
