+++ app/modules/form-place/components/validation/form-place-validation.component.ts
@@ -4,7 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 
 import { LoadingIndicator } from "nativescript-loading-indicator";
 
-import { PlaceSearchService } from '../../../../shared/place/search/place-search.service';
+import { PlaceSearchService } from '../../../../modules/google-sdk/shared/place/search/place-search.service';
 import providePlaceStorageService from '../../../../shared/place/storage/provide-service';
 import { PlaceStorageService } from '../../../../shared/place/storage/place-storage.service';
 import { CONTEXT_VALUES } from '../../../../shared/place/place';
