+++ app/modules/google-sdk/shared/place/search/place-search.service.ts
