+++ app/modules/google-sdk/shared/place/search/provide-service.ts
@@ -2,7 +2,7 @@ import {Http} from "@angular/http";
 
 import { PlaceSearchService } from './place-search.service';
 import { PlaceSearchMockService } from './place-search.mock.service';
-import { Config } from '../../config';
+import { Config } from '../../../../../shared/config';
 
 export default () => {
   return { provide: PlaceSearchService, useFactory: (http:Http) => Config.mockSearch ? new PlaceSearchMockService(http) : new PlaceSearchService(http), deps: [Http] }
