--- app/modules/search-place/components/search-place-list.html
@@ -1,9 +1,0 @@
-<StackLayout class="place-container">
-  <ListView [items]="places" separatorColor="transparent" class="list-group" height="100%" (itemTap)="onItemTap($event)">
-    <ng-template let-item="item" let-i="index" let-odd="odd" let-even="even">
-      <StackLayout class="search-list-item-container">
-        <Label [text]="item.name"></Label>
-      </StackLayout>
-    </ng-template>
-  </ListView>
-</StackLayout>
